--- frontend/lib/src/components/shared/ElementProvider/index.tsx
@@ -1,44 +1,0 @@
-/**
- * Copyright (c) Streamlit Inc. (2018-2022) Snowflake Inc. (2022-2024)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import React, { FC, PropsWithChildren } from "react"
-
-import { ToolbarRendererContextProvider } from "@streamlit/lib/src/components/shared/Toolbar/SharedToolbar"
-import { WidgetLabelRenderer } from "@streamlit/lib/src/components/widgets/BaseWidget/SharedWidgetLabel"
-import { WidgetFullscreenWrapper } from "@streamlit/lib/src/components/shared/WidgetFullscreenWrapper/WidgetFullscreenWrapper"
-
-export type ElementProviderProps = PropsWithChildren<{
-  width: number
-}>
-
-/**
- * Provides all of the necessary contexts for shared concepts. to render a
- * Streamlit element. More contexts can be added here as needed.
- */
-export const ElementProvider: FC<ElementProviderProps> = ({
-  children,
-  width,
-}) => {
-  return (
-    <WidgetFullscreenWrapper width={width}>
-      <WidgetLabelRenderer>
-        <ToolbarRendererContextProvider>
-          {children}
-        </ToolbarRendererContextProvider>
-      </WidgetLabelRenderer>
-    </WidgetFullscreenWrapper>
-  )
-}
